Move taxcalc-related functions from utils.py to taxcalc.py.
